retriever-agent/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
pinecone-client
